update 1 - 02.07.2025

+++ controllers/PostController.js
@@ -284,14 +284,23 @@ class PostController {
         try {
             const { game_id } = req.params;
             const game = await GamesModel.findById(game_id);
+            const user = req.user;
             const locale = req.cookies['locale'] || 'en';
 
             if (game.game_online.online > 0){
                 const errorMsg = locale === 'en' ? 'You cannot Delete a game that has players in it.' : 'Вы не можете Удалить игру, в котором есть игроки.';
                 return res.redirect(`/error?code=409&message=${encodeURIComponent(errorMsg)}`);
             }
+            const userInfo = await UsersModel.findById(user.id);
+
+            const notFound = !userInfo.myGames.some(g => g.gameId.toString() === game_id.toString());
+
+            if (notFound) {
+                const errorMsg = locale === 'en' ? 'Game not found.' : 'Игра не найдена.';
+                return res.redirect(`/error?message=${encodeURIComponent(errorMsg)}`);
+            }
             await GamesModel.findByIdAndDelete(game_id);
-            return res.redirect('/my-games');
+            return res.status(200).json('Игра успешно удалена!');
         } catch (err) {
             console.error(err);
             res.status(500).json({ error: err.message });
